Add create and search user functionality, api auth

+++ FCCodingChallenge.API/Controllers/UserController.cs
@@ -1,6 +1,7 @@
 ﻿using FCCodingChallenge.API.Data.Models;
 using FCCodingChallenge.API.Data.ViewModels;
 using FCCodingChallenge.API.Services;
+using FCCodingChallenge.API.Utilities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
@@ -15,8 +16,8 @@ namespace FCCodingChallenge.API.Controllers
 
         private readonly IUserService _userService;
         private readonly IUserRoleService _userRoleService;
-        public UserController(IUserService userService, IUserRoleService userRoleService, RemoteDetails remoteDetails, ILoggerManager loggerManager)
-          : base(remoteDetails, loggerManager)
+        public UserController(IUserService userService, IUserRoleService userRoleService, IHttpContextAccessor httpContext, RemoteDetails remoteDetails, ILoggerManager loggerManager)
+          : base(remoteDetails, httpContext, loggerManager)
         {
             _userService = userService;
             _userRoleService = userRoleService;
@@ -24,18 +25,26 @@ namespace FCCodingChallenge.API.Controllers
 
         [HttpGet]
         [SwaggerOperation(Summary = "Get User by Email")]
-        [Route("user/{email}")]
+        [Route("user/email/{email}")]
         public async Task<IActionResult> GetUser(string email)
         {
+            var auth = ChannelService.AuthorizeChannel(_remoteDetails.ApiKey);
+            if(!auth)
+                return Unauthorized();
+
             var resp = this.CustomResponse(await _userService.GetUser(email));
             return resp;
         }
 
         [HttpGet]
         [SwaggerOperation(Summary = "Get User by Phone")]
-        [Route("user/{phone}")]
+        [Route("user/phone/{phone}")]
         public async Task<IActionResult> GetUserByPhone(string phone)
         {
+            var auth = ChannelService.AuthorizeChannel(_remoteDetails.ApiKey);
+            if (!auth)
+                return Unauthorized();
+
             var resp = this.CustomResponse(await _userService.GetUserByPhone(phone));
             return resp;
         }
@@ -45,6 +54,10 @@ namespace FCCodingChallenge.API.Controllers
         [Route("user")]
         public async Task<IActionResult> AddUser(UserVM userRoleRequest)
         {
+            var auth = ChannelService.AuthorizeChannel(_remoteDetails.ApiKey);
+            if (!auth)
+                return Unauthorized();
+
             var resp = this.CustomResponse(await _userRoleService.AddUserRole(_remoteDetails, userRoleRequest));
             return resp;
         }
@@ -54,6 +67,10 @@ namespace FCCodingChallenge.API.Controllers
         [Route("user/{email}")]
         public async Task<IActionResult> DeleteUser(string email)
         {
+            var auth = ChannelService.AuthorizeChannel(_remoteDetails.ApiKey);
+            if (!auth)
+                return Unauthorized();
+
             var resp = this.CustomResponse(await _userRoleService.DeleteUserRole(_remoteDetails, email));
             return resp;
         }
@@ -63,6 +80,10 @@ namespace FCCodingChallenge.API.Controllers
         [Route("user")]
         public async Task<IActionResult> UpdateUser(UserVM userVM)
         {
+            var auth = ChannelService.AuthorizeChannel(_remoteDetails.ApiKey);
+            if (!auth)
+                return Unauthorized();
+
             var resp = this.CustomResponse(await _userService.UpdateUser(_remoteDetails, userVM));
             return resp;
         }
